third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFESpotLightElement-dom-pointsAtY-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'pointsAtY' attribute of the SVGFESpotLightElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS spotLightElement.getAttribute('pointsAtY') is "-30"
PASS spotLightElement.getAttribute('pointsAtY') is "70"
PASS successfullyParsed is true

TEST COMPLETE
